Actualizada libreria Time

+++ Time/Time.cpp
@@ -155,10 +155,10 @@ void breakTime(time_t timeInput, tmElements_t &tm){
 
   uint8_t year;
   uint8_t month, monthLength;
-  uint64_t time;
+  uint32_t time;
   unsigned long days;
 
-  time = (uint64_t)timeInput;
+  time = (uint32_t)timeInput;
   tm.Second = time % 60;
   time /= 60; // now it is minutes
   tm.Minute = time % 60;
@@ -207,7 +207,7 @@ time_t makeTime(const tmElements_t &tm){
 // previous version used full four digit year (or digits since 2000),i.e. 2009 was 2009 or 9
   
   int i;
-  uint64_t seconds;
+  uint32_t seconds;
 
   // seconds from 1970 till 1 jan 00:00:00 of the given year
   seconds= tm.Year*(SECS_PER_DAY * 365);
@@ -234,9 +234,9 @@ time_t makeTime(const tmElements_t &tm){
 /*=====================================================*/	
 /* Low level system time functions  */
 
-static uint64_t sysTime = 0;
-static uint64_t prevMillis = 0;
-static uint64_t nextSyncTime = 0;
+static uint32_t sysTime = 0;
+static uint32_t prevMillis = 0;
+static uint32_t nextSyncTime = 0;
 static timeStatus_t Status = timeNotSet;
 
 getExternalTime getTimePtr;  // pointer to external sync function
@@ -277,8 +277,8 @@ void setTime(time_t t) {
    sysUnsyncedTime = t;   // store the time of the first call to set a valid Time   
 #endif
 
-  sysTime = (uint64_t)t;  
-  nextSyncTime = (uint64_t)t + syncInterval;
+  sysTime = (uint32_t)t;  
+  nextSyncTime = (uint32_t)t + syncInterval;
   Status = timeSet;
   prevMillis = millis();  // restart counting from now (thanks to Korman for this fix)
 } 
@@ -316,6 +316,6 @@ void setSyncProvider( getExternalTime getTimeFunction){
 }
 
 void setSyncInterval(time_t interval){ // set the number of seconds between re-sync
-  syncInterval = (uint64_t)interval;
+  syncInterval = (uint32_t)interval;
   nextSyncTime = sysTime + syncInterval;
 }
